LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/unhandled-promise-rejections/allow-crossorigin-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: 42

PASS Promise rejection event should be received for the cross-origin CORS script
